// src/generatedModels/NotFoundError.generated.ts
/* eslint-disable */
/**
 * This file was automatically generated by json-schema-to-typescript.
 * DO NOT MODIFY IT BY HAND. Instead, modify the source JSONSchema file.
 *
 * =>  @see ./src/schemas/NotFoundError.yml
 *
 * and run "npm run schema-to-interface" or "npm run generate" to regenerate this file.
 */

export interface NotFoundError {
  error?: {
    msg?: string;
  };
}